requirements.txt
pika==1.3.1
grpcio==1.59.3
grpcio-tools==1.59.3
pydub==0.25.1
SpeechRecognition==3.10.0

# bias-requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
transformers==4.35.2

# flask-requirements.txt
flask==3.0.0
pika==1.3.1